Add IconView for displaying author images and update AuthorInitialIcon usage in CommitRowView
- Introduced IconView to handle user avatars with gravatar support.
- Changed static method parameter type from Int to CGFloat for size.
- Updated CommitRowView to utilize IconView instead of custom code for displaying author initials.

+++ GitClient/Extensions/URL+.swift
@@ -14,7 +14,7 @@ extension URL {
         return URL(fileURLWithPath: srcroot).appending(path: "TestFixtures").appending(path: "SyntaxHighlight")
     }
 
-    static func gravater(email: String, size: Int=80) -> URL? {
+    static func gravater(email: String, size: CGFloat=80) -> URL? {
         guard let data = email.trimmingCharacters(in: .whitespacesAndNewlines).lowercased().data(using: .utf8) else {
             return nil
         }
